requirements.txt
python-multipart
openai
twilio
python-dotenv
pydub
requests
fastapi
uvicorn
pyngrok
pydantic
colorama
